[CELEBORN-502][REFACTOR] Merge GetBlacklistResponse to HeartbeatFromApplication
