+++ client/src/main/scala/org/apache/celeborn/client/ChangePartitionManager.scala
@@ -207,7 +207,7 @@ class ChangePartitionManager(
     // Blacklist all failed workers
     if (changePartitions.exists(_.causes.isDefined)) {
       changePartitions.filter(_.causes.isDefined).foreach { changePartition =>
-        lifecycleManager.blacklistPartition(
+        lifecycleManager.blacklistManager.blacklistPartition(
           shuffleId,
           changePartition.oldPartition,
           changePartition.causes.get)
@@ -252,7 +252,7 @@ class ChangePartitionManager(
         .workerSnapshots(shuffleId)
         .keySet()
         .asScala
-        .filter(w => !lifecycleManager.blacklist.keySet().contains(w))
+        .filter(w => !lifecycleManager.blacklistManager.blacklist.keySet().contains(w))
         .toList
     if (candidates.size < 1 || (pushReplicateEnabled && candidates.size < 2)) {
       logError("[Update partition] failed for not enough candidates for revive.")
