+++ common/src/main/proto/TransportMessages.proto
@@ -39,8 +39,8 @@ enum MessageType {
   APPLICATION_LOST = 18;
   APPLICATION_LOST_RESPONSE = 19;
   HEARTBEAT_FROM_APPLICATION = 20;
-  GET_BLACKLIST = 21;
-  GET_BLACKLIST_RESPONSE = 22;
+  HEARTBEAT_FROM_APPLICATION_RESPONSE = 21;
+  // Reserve 22
   CHECK_QUOTA = 23;
   CHECK_QUOTA_RESPONSE = 24;
   REPORT_WORKER_FAILURE = 25;
@@ -282,13 +282,10 @@ message PbHeartbeatFromApplication {
   int64 totalWritten = 2;
   int64 fileCount = 3 ;
   string requestId = 4;
+  repeated PbWorkerInfo localBlackList = 5;
 }
 
-message PbGetBlacklist {
-  repeated PbWorkerInfo localBlackList = 1;
-}
-
-message PbGetBlacklistResponse {
+message PbHeartbeatFromApplicationResponse {
   int32 status = 1;
   repeated PbWorkerInfo blacklist = 2;
   repeated PbWorkerInfo unknownWorkers = 3;
